Keyboard Input
Added Keyboard input to the Window. All key states are stored in one of 3 u64. One for presses, one for being held down, and one for being released. Using the & operator with one of the previously defined enums lets you know if the key is in any of the three states. Also with this current method keyboard inputs can be simulated in the program at runtime.

Still only works on Windows(os) and I don't plan to implement others. Maybe very far in the future I will.